Created edit inventory page
The current issue is that the edit page needs the resource to populate the fields in the form, but it cannot get that information before rendering.  I am going to work around this by converting the page to a modal on the inventory page so the properties of an economicResource can be directly passed without re-querying.  This will also help with permission issues.